+++ client/src/pages/EvaluateTeam/index.js
@@ -2,28 +2,19 @@ import React, { useState, useEffect } from 'react'
 import './styles.css'
 import TeamList from '../../components/TeamList'
 import api from '../../services/api'
-import team0 from '../../assets/team1.svg'
-import team1 from '../../assets/team2.svg'
-import team2 from '../../assets/team3.svg'
-import team3 from '../../assets/team4.svg'
+import getRandomSvg from '../../utils/getRandomSvg'
 
 export default function EvaluateTeam() {
-
   const [teams, setTeams] = useState([])
 
   useEffect(() => {
     handleData()
   }, [])
 
-  function getRandomSvg() {
-    let vector = [team0, team1, team2, team3]
-    return vector[Math.floor(Math.random() * 4)]
-  }
-
-  async function handleData(){
+  async function handleData() {
     const data = (await api.get('/teams-complete')).data
     data.map((team) => {
-       team['avatar'] = getRandomSvg()
+      team['avatar'] = getRandomSvg('team')
     })
 
     await setTeams(data)
